@@ -1,6 +1,6 @@
-module go.source.hueristiq.com/retrier
+module github.com/hueristiq/hq-go-retrier
 
-go 1.23.3
+go 1.24.2
 
 require github.com/stretchr/testify v1.10.0
 
